refactor: consolidate DFIQ archive validation into tests/tools/validate_dfiq_archive.py
Drops the standalone tools/ package - the validate() function and its
tests now live together in a single file. magika is installed as a CI-only
step instead of a project dependency, since it's only needed for this
check.
